+++ changelog.md
@@ -1,6 +1,7 @@
 # v1.0.2
 - Minor changes
 - Fixed **Preview Music mid-song** to work on MacOS
+- Fixed issues with preview music still playing when starting the level
 # v1.0.1
 - Reworking the retry logic for built-in tracks
 - Attempted to fix the audio playback issues for mobile users